Milestone 4. Task 3/requirements.txt
absl-py==2.1.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth-oauthlib==1.0.0
google-auth==2.26.2
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
ml-dtypes==0.2.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pip==22.3.1
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1-modules==0.3.0
pyasn1==0.5.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests-oauthlib==1.3.1
requests==2.31.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.1
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==65.5.1
six==1.16.0
smmap==5.0.1
tensorboard-data-server==0.7.2
tensorboard==2.14.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflow==2.14.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wandb==0.16.2
werkzeug==3.0.1
wheel==0.38.4
wrapt==1.14.1
zipp==3.17.0
